requirements.txt
# GraphRAG Query Router
# Core graph DB
neo4j==5.23.1

# Embeddings
sentence-transformers==3.3.1
torch>=2.0.0

# Vector search
faiss-cpu==1.9.0
numpy>=1.24.0

# Graph algorithms (community detection)
networkx==3.4.2
python-louvain==0.16
igraph==0.11.8
leidenalg==0.10.2

# API
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4

# Utilities
python-dotenv==1.0.1
loguru==0.7.3
tqdm==4.67.1
scikit-learn==1.5.2
